requirements.txt
fastapi[standard]==0.115.0
redis==5.2.0

psi-datahub==1.0.1
numpy==2.1.3
h5py==3.12.1
requests==2.32.3
pyepics==3.5.7
pyzmq==26.2.0
cbor2==5.6.5
python-dateutil==2.9.0.post0
bitshuffle==0.5.2
pytz==2024.2
pandas==2.2.3
matplotlib==3.9.3
psutil==6.1.0